requirements.txt
PyYAML
Markdown
uuid